fix circular types
