+++ packages/core/src/types.ts
@@ -1,10 +1,28 @@
-export type Alignment = 'start' | 'end';
-export type Side = 'top' | 'right' | 'bottom' | 'left';
-export type AlignedPlacement = `${Side}-${Alignment}`;
-export type Placement = Side | AlignedPlacement;
-export type Strategy = 'absolute' | 'fixed';
-export type Axis = 'x' | 'y';
-export type Length = 'width' | 'height';
+import {
+  Axis,
+  ClientRectObject,
+  ElementRects,
+  Length,
+  Placement,
+  Rect,
+  SideObject,
+  Strategy,
+} from '@floating-ui/utils';
+export type {
+  AlignedPlacement,
+  Alignment,
+  Axis,
+  ClientRectObject,
+  ElementRects,
+  Length,
+  Padding,
+  Placement,
+  Rect,
+  Side,
+  SideObject,
+  Strategy,
+  VirtualElement,
+} from '@floating-ui/utils';
 
 type Promisable<T> = T | Promise<T>;
 
@@ -47,8 +65,6 @@ export interface Platform {
 }
 
 export type Coords = {[key in Axis]: number};
-
-export type SideObject = {[key in Side]: number};
 
 export interface MiddlewareData {
   [key: string]: any;
@@ -140,21 +156,6 @@ export type Middleware = {
 
 export type Dimensions = {[key in Length]: number};
 
-export type Rect = Coords & Dimensions;
-
-export interface ElementRects {
-  reference: Rect;
-  floating: Rect;
-}
-
-/**
- * Custom positioning reference element.
- * @see https://floating-ui.com/docs/virtual-elements
- */
-export type VirtualElement = {
-  getBoundingClientRect(): ClientRectObject;
-  contextElement?: any;
-};
 export type ReferenceElement = any;
 export type FloatingElement = any;
 
@@ -177,8 +178,6 @@ export interface MiddlewareState extends Coords {
  */
 export type MiddlewareArguments = MiddlewareState;
 
-export type ClientRectObject = Rect & SideObject;
-export type Padding = number | Partial<SideObject>;
 export type Boundary = any;
 export type RootBoundary = 'viewport' | 'document' | Rect;
 export type ElementContext = 'reference' | 'floating';
