+++ packages/utils/index.d.ts
@@ -1,1 +1,1 @@
-export * from './src';
+export * from './src/types';
