+++ packages/utils/package.json
@@ -69,8 +69,5 @@
     "menu",
     "popup",
     "positioning"
-  ],
-  "dependencies": {
-    "@floating-ui/core": "^0.1.0"
-  }
+  ]
 }
